cv-service/requirements.txt
opencv-python-headless==4.9.0.80
ultralytics==8.1.0
numpy>=1.24.0
requests>=2.31.0
supabase>=2.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
Pillow>=10.0.0
